Fix the length issue that is related to Metal compiler.

+++ lib/nnc/mfa/kernels/Dequantize8iRowwiseXKernel.cpp
@@ -403,7 +403,7 @@ inline char4 q8_char4(const int4 v)
   return char4((char)v.x, (char)v.y, (char)v.z, (char)v.w);
 }
 
-inline void store_q8_scalar(device uchar* destination, const ulong destination_offset, const uint col_base, thread int* q8)
+inline void store_q8_scalar(device uchar* destination, const uint destination_offset, const uint col_base, thread int* q8)
 {
   device int8_t* destination_q = reinterpret_cast<device int8_t*>(destination);
   for (uint j = 0; j < group_size; ++j) {
@@ -491,7 +491,7 @@ inline int4 q5_tail_values(const ulong lo, const uint hi, const int m, const int
     (int)((hi >> 11u) & 31u)) - int4(16)) * int4(m) + int4(b);
 }
 
-inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const ulong destination_offset, const uint col_base)
+inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const uint destination_offset, const uint col_base)
 {
   device const uchar* p = source + group_index * 11ul;
   const ulong lo =
@@ -534,7 +534,7 @@ inline int4 q6_values_i4(const ulong payload, const uint lane, const int m, cons
     q6_signed_value((uint)((payload >> ((lane + 3u) * 6u)) & 63ul))) * int4(m) + int4(b);
 }
 
-inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const ulong destination_offset, const uint col_base)
+inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const uint destination_offset, const uint col_base)
 {
   const ulong payload = packed52_payload(source, group_index);
   const int m = (int)((payload >> 48u) & 3ul) + 1;
@@ -564,7 +564,7 @@ inline int4 q4_values(device const uchar* p, const uint offset, const int m, con
     (int)(q1 >> 4)) - int4(8)) * int4(m) + int4(b);
 }
 
-inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const ulong destination_offset, const uint col_base)
+inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const uint destination_offset, const uint col_base)
 {
   device const uchar* p = source + group_index * 9ul;
   const int m = (int)(p[8] & 15u) + 1;
@@ -594,7 +594,7 @@ inline int4 q3_values_i4(const uint q, const int m, const int b)
     (int)((q >> 9u) & 7u)) - int4(4)) * int4(m) + int4(b);
 }
 
-inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const ulong destination_offset, const uint col_base)
+inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const uint destination_offset, const uint col_base)
 {
   device const uchar* p = source + group_index * 7ul;
   const uint lo =
@@ -633,7 +633,7 @@ inline int4 q2_values_i4(const uint q, const uint lane, const int m, const int z
     (int)((q >> ((lane + 3u) * 2u)) & 3u)) * int4(m) - int4(z);
 }
 
-inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const ulong destination_offset, const uint col_base)
+inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const uint destination_offset, const uint col_base)
 {
   const ulong payload = packed42_payload(source, group_index);
   const uint q = (uint)payload;
@@ -655,7 +655,7 @@ inline void decode_store_group(device const uchar* source, const ulong group_ind
 )";
 	} else if (format == CCV_NNC_QX_8I_ROWWISE_IQ2_XXS) {
 		shader += R"(
-inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const ulong destination_offset, const uint col_base)
+inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const uint destination_offset, const uint col_base)
 {
   device const uchar* p = source + group_index * 8ul;
   const uint sign_codes = (uint)p[4] | ((uint)p[5] << 8) | ((uint)p[6] << 16) | (((uint)p[7] & 15u) << 24);
@@ -677,7 +677,7 @@ inline void decode_store_group(device const uchar* source, const ulong group_ind
 )";
 	} else if (format == CCV_NNC_QX_8I_ROWWISE_IQ2_S) {
 		shader += R"(
-inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const ulong destination_offset, const uint col_base)
+inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const uint destination_offset, const uint col_base)
 {
   const ulong payload = packed42_payload(source, group_index);
   const uint grid0 = (uint)(payload & 1023u);
@@ -700,7 +700,7 @@ inline void decode_store_group(device const uchar* source, const ulong group_ind
 )";
 	} else if (format == CCV_NNC_QX_8I_ROWWISE_IQ2_XS) {
 		shader += R"(
-inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const ulong destination_offset, const uint col_base)
+inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const uint destination_offset, const uint col_base)
 {
   const uint payload = packed21_payload(source, group_index);
   const uint grid0 = payload & 511u;
@@ -720,7 +720,7 @@ inline void decode_store_group(device const uchar* source, const ulong group_ind
 )";
 	} else if (format == CCV_NNC_QX_8I_ROWWISE_IQ3_S) {
 		shader += R"(
-inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const ulong destination_offset, const uint col_base)
+inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const uint destination_offset, const uint col_base)
 {
   device const uchar* p = source + group_index * 7ul;
   const uint grid0 = (uint)p[0] | (((uint)p[1] & 1u) << 8);
@@ -745,7 +745,7 @@ inline void decode_store_group(device const uchar* source, const ulong group_ind
 )";
 	} else if (format == CCV_NNC_QX_8I_ROWWISE_IQ3_XXS) {
 		shader += R"(
-inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const ulong destination_offset, const uint col_base)
+inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const uint destination_offset, const uint col_base)
 {
   const uint payload = packed28_payload(source, group_index);
   const uint grid0 = payload & 255u;
@@ -766,7 +766,7 @@ inline void decode_store_group(device const uchar* source, const ulong group_ind
 )";
 	} else {
 		shader += R"(
-inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const ulong destination_offset, const uint col_base)
+inline void decode_store_group(device const uchar* source, const ulong group_index, device uchar* destination, const uint destination_offset, const uint col_base)
 {
 )";
 		if (format == CCV_NNC_QX_8I_ROWWISE_IQ2_XXS)
@@ -798,7 +798,7 @@ kernel void dequantize_8i_rowwise_x(
     const uint row = x / groups_per_row;
     const uint group = x - row * groups_per_row;
     const uint col_base = group * group_size;
-    decode_store_group(source, x, destination, (ulong)row * row_length + col_base, col_base);
+    decode_store_group(source, x, destination, row * row_length + col_base, col_base);
   }
   if (x < scale_bytes)
     output_scales[x] = input_scales[x];
@@ -825,12 +825,12 @@ kernel void dequantize_8i_rowwise_x_selected(
     const uint row = x / groups_per_row;
     const uint group = x - row * groups_per_row;
     const uint col_base = group * group_size;
-    const ulong destination_row = (ulong)expert * rows_per_expert + row;
+    const uint destination_row = expert * rows_per_expert + row;
     decode_store_group(source, (ulong)expert * groups_per_expert + x, destination, destination_row * row_length + col_base, col_base);
   }
   if (x < scale_bytes_per_expert) {
-    output_scales[(ulong)expert * scale_bytes_per_expert + x] =
-      input_scales[(ulong)expert * scale_bytes_per_expert + x];
+    output_scales[expert * scale_bytes_per_expert + x] =
+      input_scales[expert * scale_bytes_per_expert + x];
   }
 }
